Map interpolation is changed: choroidMap.m: Now uses the "natural" method for scatterInterpolant, and uses "nearest" for extrapolation.
The segmentation of the BM has been changed. The function getRetinaAndRPE.m has been renamed to getRetinaAndBM.m. The function find the RPE as the centroid for each A-scan, and then uses a graph search weighted as the negative intensity slope to estimate the location of the BM. Then, this trace is used to estimate the distance between the center of the RPE and the negative slope. Finally the bottom of the RPE is estimated as the center plus the distance computed before. This is because the center of the RPE is more robust than the trace of the slope. At the end, the convex hull is used as before. Now the BM is better estimated, since it is a bit below the RPE, and not on top of it.